refactor of some table component
