+++ src/WebUI/Common/KubeCardWithTable.scss
